moved my now reuseable fetchData call into a data folder
